refactered the code to return number of tokens of a single message
