+++ src/helpers/token_counter.py
@@ -1,0 +1,74 @@
+"""
+https://cookbook.openai.com/examples/how_to_count_tokens_with_tiktoken
+"""
+import tiktoken
+
+def tokens_per_model(model="gpt-3.5-turbo-0613") -> tuple[int, int]:
+    if model in {
+        "gpt-3.5-turbo-0613",
+        "gpt-3.5-turbo-16k-0613",
+        "gpt-4-0314",
+        "gpt-4-32k-0314",
+        "gpt-4-0613",
+        "gpt-4-32k-0613",
+    }:
+        tokens_per_message = 3
+        tokens_per_name = 1
+    elif model == "gpt-3.5-turbo-0301":
+        tokens_per_message = 4  # every message follows <|start|>{role/name}\n{content}<|end|>\n
+        tokens_per_name = -1  # if there's a name, the role is omitted
+    elif "gpt-3.5-turbo" in model:
+        # print("Warning: gpt-3.5-turbo may update over time. Returning num tokens assuming gpt-3.5-turbo-0613.")
+        return tokens_per_model(model="gpt-3.5-turbo-0613")
+    elif "gpt-4" in model:
+        # print("Warning: gpt-4 may update over time. Returning num tokens assuming gpt-4-0613.")
+        return tokens_per_model(model="gpt-4-0613")
+    else:
+        raise NotImplementedError(
+            f"""tokens_per_model() is not implemented for model {model}. See https://github.com/openai/openai-python/blob/main/chatml.md for information on how messages are converted to tokens."""
+        )
+
+    return (tokens_per_message, tokens_per_name)
+
+def get_encoding_for_model(model="gpt-3.5-turbo-0613"):
+    try:
+        encoding = tiktoken.encoding_for_model(model)
+    except KeyError:
+        print("Warning: model not found. Using cl100k_base encoding.")
+        encoding = tiktoken.get_encoding("cl100k_base")
+
+    return encoding
+
+def num_tokens_from_messages(messages, model="gpt-3.5-turbo-0613"):
+    """
+    Return the number of tokens used by a list of messages.
+    """
+    encoding = get_encoding_for_model(model=model)
+    (tokens_per_message, tokens_per_name) = tokens_per_model(model=model)
+    num_tokens = 0
+    for message in messages:
+        num_tokens += tokens_per_message
+        if not hasattr(message, 'items'):
+            continue
+        for key, value in message.items():
+            num_tokens += len(encoding.encode(value))
+            if key == "name":
+                num_tokens += tokens_per_name
+    num_tokens += 3  # every reply is primed with <|start|>assistant<|message|>
+
+    return num_tokens
+
+def num_tokens_from_message(message, model="gpt-3.5-turbo-0613"):
+    """
+    Return the number of tokens used by a message.
+    """
+    encoding = get_encoding_for_model(model=model)
+
+    (tokens_per_message, tokens_per_name) = tokens_per_model(model=model)
+    num_tokens = 0
+    num_tokens += tokens_per_message
+    num_tokens += len(encoding.encode(message))
+    num_tokens += tokens_per_name
+    num_tokens += 3  # every reply is primed with <|start|>assistant<|message|>
+
+    return num_tokens
